moved util
